+++ distributaur/example/worker.py
@@ -3,6 +3,9 @@ import sys
 
 sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), "./"))
 
-from distributaur.example.local import distributaur
+from distributaur.example.shared import distributaur, example_function
 
 celery = distributaur.app
+
+if __name__ == "main":
+    distributaur.register_function(example_function)
