+++ distributaur/tests/test.py
@@ -1,3 +1,4 @@
+from io import StringIO
 import json
 import pytest
 import subprocess
@@ -9,7 +10,7 @@ from unittest.mock import MagicMock, patch
 
 from huggingface_hub import HfApi
 
-sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), "../"))
+sys.path.append(os.path.join(os.path.dirname(os.path.abspath(__file__)), "../../"))
 
 from distributaur.distributaur import Distributaur
 from distributaur.tests.worker import example_function
@@ -75,34 +76,35 @@ def test_execute_function():
     print("Task execution test passed")
 
 
-def test_worker_task_execution():
-    distributaur = Distributaur()
+# def test_worker_task_execution():
+#     distributaur = Distributaur()
 
-    distributaur.register_function(example_function)
+#     distributaur.register_function(example_function)
 
-    worker_cmd = [
-        "celery",
-        "-A",
-        "distributaur.tests.worker",
-        "worker",
-        "--loglevel=info",
-    ]
-    print("worker_cmd")
-    print(worker_cmd)
-    worker_process = subprocess.Popen(worker_cmd)
+#     worker_cmd = [
+#         "celery",
+#         "-A",
+#         "distributaur.tests.worker",
+#         "worker",
+#         "--loglevel=info",
+#     ]
+#     print("worker_cmd")
+#     print(worker_cmd)
+#     worker_process = subprocess.Popen(worker_cmd)
 
-    time.sleep(5)
+#     time.sleep(2)
 
-    task_params = {"arg1": 10, "arg2": 20}
-    task = distributaur.execute_function("example_function", task_params)
-    result = task.get(timeout=3)
+#     task_params = {"arg1": 10, "arg2": 20}
+#     print("executing task")
+#     task = distributaur.execute_function("example_function", task_params)
+#     result = task.get(timeout=30)
 
-    assert result == "Result: arg1=30"
+#     assert result == "Result: arg1=30"
 
-    worker_process.terminate()
-    worker_process.wait()
+#     worker_process.terminate()
+#     worker_process.wait()
 
-    print("Worker task execution test passed")
+#     print("Worker task execution test passed")
 
 
 def test_task_status_update():
@@ -378,3 +380,56 @@ def test_create_instance(mock_put):
     instance = distributaur.create_instance(offer_id, image)
 
     assert instance["new_contract"] == "instance1"
+
+
+def test_local_example_run():
+
+    # Capture the stdout and stderr during the execution
+    with patch("sys.stdout", new=StringIO()) as fake_out, patch(
+        "sys.stderr", new=StringIO()
+    ) as fake_err:
+        # verify that the local.py exists, ls the files if it doesnt
+        if not os.path.exists("distributaur/example/local.py"):
+            print("Local.py does not exist")
+        exec(open("distributaur/example/local.py").read())
+
+    # def test_distributed_example_function():
+    #     distributaur = Distributaur()
+    #     distributaur.register_function(example_function)
+
+    #     task_params = {"index": 0, "arg1": 5, "arg2": 10}
+    #     result = example_function(**task_params)
+
+    #     assert result == "Task 0 completed. Result (5 + 10): 15"
+
+# @patch("distributaur.distributaur.Distributaur.rent_nodes")
+# @patch("distributaur.distributaur.Distributaur.terminate_nodes")
+# def test_distributed_example_run(mock_rent_nodes, mock_terminate_nodes):
+#     # Mock the rent_nodes method to return a list of rented nodes
+#     mock_rent_nodes.return_value = [{"instance_id": "instance1"}]
+
+#     # Run the main block of code from distributed.py
+#     # You can modify the code to accept command-line arguments for testing purposes
+#     # For example, you can pass the number of tasks as an argument
+#     sys.argv = ["example/distributed.py", "--tasks", "3"]
+
+#     # Capture the stdout and stderr during the execution
+#     with patch("sys.stdout", new=StringIO()) as fake_out, patch(
+#         "sys.stderr", new=StringIO()
+#     ) as fake_err:
+#         exec(open("example/distributed.py").read())
+
+#     # Assert the expected output or behavior
+#     # For example, check if the expected number of tasks were submitted
+#     assert "TOTAL NODES AVAILABLE:" in fake_out.getvalue()
+#     assert "TOTAL RENTED NODES:" in fake_out.getvalue()
+#     assert "Submitting tasks..." in fake_out.getvalue()
+#     assert (
+#         "Tasks submitted to queue. Waiting for tasks to complete..."
+#         in fake_out.getvalue()
+#     )
+#     assert "Worker process terminated." in fake_out.getvalue()
+#     assert "Example completed." in fake_out.getvalue()
+
+#     # Verify that the rented nodes were terminated
+#     mock_terminate_nodes.assert_called_once()
