+++ distributaur/tests/worker.py
@@ -3,12 +3,13 @@ from distributaur.distributaur import Distributaur
 distributaur = Distributaur()
 
 
-# Define and register the example_function
+# Define and register the test_function
 def example_function(arg1, arg2):
     return f"Result: arg1={arg1+arg2}"
 
 
+celery = distributaur.app
+
+
 if __name__ == "__main__":
     distributaur.register_function(example_function)
-
-    celery = distributaur.app
